Muutoksia

+++ ListAndLoopsExample/Coffee.cs
@@ -9,9 +9,10 @@ namespace ListAndLoopsExample
         public enum Roast {light=1, lightmedium, medium, darkmedium, dark}
 
         public string brand;
-        public decimal price;
+        public double price;
         public Roast roast;
         public string importer;
+        
         public Coffee()
         {
 
@@ -22,7 +23,7 @@ namespace ListAndLoopsExample
             this.brand = brand;
         }
 
-        public Coffee(string brand, decimal price, Roast roast, string importer = "default")
+        public Coffee(string brand, double price, Roast roast, string importer = "default")
         {
             this.brand = brand;
             this.price = price;
